requirements.txt
alembic==1.14.0
blinker==1.9.0
cache==1.0.3
cachelib==0.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
Flask-Caching==2.3.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.5
Mako==1.3.8
MarkupSafe==3.0.2
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
